fix : mypage 코드 정리 #27
