+++ src/components/mypage/reservedTickets/PurchaseTicket.stories.tsx
