+++ src/components/mypage/slides/Slider.stories.tsx
