+++ src/pages/mypage/OrderList.tsx
@@ -6,19 +6,12 @@ import useGetOrders from '../../hooks/queries/useGetOrders';
 import { IOrder } from '../../apis/type/order';
 import OrderedTicketSkeleton from '../../components/skeleton/OrderedTicketSkeleton';
 import useGetUserInfo from '../../hooks/queries/useGetUserInfo';
+import { convertDate } from '../../utils/convertDate';
 
 const OrderList = () => {
   const { status: userInfoStatus, data: userInfo } = useGetUserInfo();
   const { status, data: orders } = useGetOrders();
-  const convertDate = (date: string) => {
-    const week = new Array('일', '월', '화', '수', '목', '금', '토');
 
-    const Year = date.substring(2, 4);
-    const Month = date.substring(5, 7);
-    const Day = date.substring(8, 10);
-    const DayKR = week[new Date(`2022-${Month}-${Day}`).getDay()];
-    return { Year, Month, Day, DayKR };
-  };
   return (
     <>
       <AppBar label={'마이페이지'} />
